rename: PyStructure → HexMaps (imports, strings, CLI, docs, pyproject.toml)
