+++ hexmaps/__init__.py
@@ -1,5 +1,5 @@
 """
-PyStructure Pipeline: homogenize and analyze multi-wavelength astronomical datasets.
+HexMaps Pipeline: homogenize and analyze multi-wavelength astronomical datasets.
 """
 
 __author__ = "J. den Brok & L. Neumann"
@@ -7,7 +7,7 @@ __version__ = "4.3.0"
 __email__ = "jadenbrok@mpia.de & lukas.neumann@eso.org"
 __credits__ = ["M. Jimenez-Donaire", "E. Rosolowsky", "A. Leroy", "I. Beslic"]
 
-from pystructurePipeline.handler_pipeline import PipelineHandler
-from pystructurePipeline.init_workdir import init_workdir
+from hexmaps.handler_pipeline import PipelineHandler
+from hexmaps.init_workdir import init_workdir
 
 __all__ = ["PipelineHandler", "init_workdir"]
